requirements.txt
fastapi
uvicorn
gtts
google-generativeai
python-multipart
